secretsegypt3-main/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==4.0.0
crispy-forms-gds==0.2.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.8
django-admin-comments==1.1.0
django-bootstrap==0.2.4
django-bootstrap-static==4.2.1
django-bootstrap4==2.2.0
django-ckeditor==5.9.0
django-configurations==2.2
django-crispy-forms==1.10.0
django-geoip2-extras==1.2.1
django-heroku==0.3.1
django-import-export==2.4.0
django-js-asset==1.2.2
django-summernote==0.8.11.6
django-taggit==1.3.0
django-timezone-field==4.0
et-xmlfile==1.0.1
filelock==3.0.12
geoip2==3.0.0
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
maxminddb==2.0.3
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.1.2
pipenv==2020.11.15
psycopg2==2.8.6
python-dateutil==2.8.1
python-env-utils==0.4.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
rope==0.18.0
sentry-sdk==0.16.2
six==1.15.0
somepackage==1.2.3
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
virtualenv==20.0.23
virtualenv-clone==0.5.4
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
